vlc_fingerprinter: use size_t for index argument in callback
It does not make sense for the passed index to be negative, nor is the
current diagnostic related to comparision of different signedness very
nice to look at.

This fixes both issues by changing the argument to size_t, making sure
that the values passed are suitable for interaction with
vlc_array-related functions.

+++ include/vlc_fingerprinter.h
@@ -77,7 +77,7 @@ struct fingerprinter_thread_t
 
     void ( *pf_enqueue ) ( struct fingerprinter_thread_t *f, fingerprint_request_t *r );
     fingerprint_request_t * ( *pf_getresults ) ( struct fingerprinter_thread_t *f );
-    void ( *pf_apply ) ( fingerprint_request_t *, int i_resultid );
+    void ( *pf_apply ) ( fingerprint_request_t *, size_t i_resultid );
 };
 typedef struct fingerprinter_thread_t fingerprinter_thread_t;
